* test: 테스트 컴파일 오류 수정

* fix: POST 요청 시 돌려주는 URI 수정

* refactor: 쿼리문 로컬 변수 명 수정

* feat: 데이터베이스에서 전체 게임 id 리스트를 꺼내는 기능 구현

* test: 테스트 오류 수정

* style: 메서드 위치 수정

* style: 메서드 명 수정

* feat: Service에 전체 게임 리스트 조회 기능 추가

* feat: Controller에 전체 게임 리스트 조회 기능 매핑

* fix: 클라이언트 페이지 수정

* refactor: 예외 처리 핸들러 분리

* feat: 클라이언트에서 404 에러를 처리하는 페이지 구현

* fix: 없는 게임 방을 직접 주소로 치고 들어가면 게임이 생성되는 문제 해결

* refactor: service 객체가 상태를 갖지 않도록 변경

* feat: 이미 존재하는 게임을 생성하려고 하면 중복이라는 메시지를 담은 Response를 던지도록 하는 기능 추가

* feat: 게임 이름과 비밀번호를 받아서 게임 생성을 요청하는 기능 추가

* fix: 게임 생성 시 게임 이름과 비밀번호를 받아서 생성하도록 변경

* fix: 게임 목록을 보여줄 때 게임 번호가 아닌 게임 이름을 보여주도록 변경

* refactor: 불필요한 닫는 html 태그 제거

* feat: 비밀번호로 인증하여 게임을 종료 및 삭제하는 기능 추가

* test: 서비스 테스트 시 가짜 DAO 객체 사용하도록 변경

* test: WebController 테스트 시 Present Layer만 스캔하도록 변경

* test: 불필요한 주석 및 어노테이션 제거

* chore: jdbc 드라이버 교체

* refactor: id 값을 long 타입에서 Long 타입으로 변경

* fix: 게임이 끝난 후에만 게임을 삭제할 수 있도록 수정

* refactor: 불필요한 enum 제거

* fix: 버튼 이름을 재시작에서 초기화로 변경

* feat: 게임 이름을 기준으로 중복을 검사하는 기능 추가

* fix: 출력 메시지 오타 수정

* test: 이름이 제대로 부여되지 않은 테스트 이름 부여

* refactor: sql parameter로 Map 대신 SqlParameterSource 사용

* test: DAO 테스트 시 불필요한 스프링 로드 제거

* refactor: 사용하지 않는 메서드 제거

* test: 에러 메시지 테스트 오류 수정

* refactor: DAO를 인터페이스 - 구현체 형태에서 구현체 형태로 변경

* refactor: 게임의 컬럼명을 간소화하고 id가 자동 생성 되도록 변경

* refactor: 에러 페이지 매핑에 스프링부트의 기능 사용

* refactor: insert 과정에서 DuplicateKeyException 발생 시 에러 메시지를 직접 담도록 변경

* refactor: game 테이블에 접근 시 모든 컬럼에 대해 직접 접근하지 않고 GameEntity를 만들어 삽입/조회하도…

* fix: 터지는 예외가 NoSuchElementException에서 EmptyResultDataAccessException으…

* feat: 올바르지 않은 값(null)이 들어 왔을 때 예외 처리 추가

* refactor: 비밀번호를 Authorization 헤더에서 커스텀 헤더에 담도록 수정

* refactor: Entity의 setter 제거

* refactor: 불필요한 try ~ catch 제거

* feat: 컨트롤러로 들어오는 값에 대한 검증 로직 추가

* refactor: 불필요한 if 문 제거

* refactor: DAO 중복 접근 제거

* test: 비밀번호 암호화 테스트 추가

* refactor: 정적 팩터리 메서드로 엔티티의 용도에 따라 생성 분리

* feat: 예상하지 못한 예외에 대한 처리 로직 추가

* test: PathVariable 타입 예외 테스트 추가

* refactor: 불필요하게 분리된 메서드 병합

* chore: 패키지 의존 순환 피하기 위해 dto 패키지 위치 이동

+++ src/main/java/chess/service/ChessService.java
@@ -1,12 +1,8 @@
 package chess.service;
 
-import chess.controller.dto.request.MoveRequest;
-import chess.controller.dto.response.ChessGameResponse;
-import chess.controller.dto.response.EndResponse;
-import chess.controller.dto.response.PieceResponse;
-import chess.controller.dto.response.StatusResponse;
 import chess.dao.GameDao;
 import chess.dao.PieceDao;
+import chess.dao.entity.GameEntity;
 import chess.domain.ChessGame;
 import chess.domain.GameState;
 import chess.domain.board.Board;
@@ -15,37 +11,70 @@ import chess.domain.board.Position;
 import chess.domain.board.Row;
 import chess.domain.board.strategy.CreateCompleteBoardStrategy;
 import chess.domain.piece.Piece;
+import chess.dto.request.CreateGameRequest;
+import chess.dto.request.MoveRequest;
+import chess.dto.response.ChessGameResponse;
+import chess.dto.response.ChessGamesResponse;
+import chess.dto.response.GameRoomResponse;
+import chess.dto.response.PieceResponse;
+import chess.dto.response.StatusResponse;
+import chess.util.PasswordEncryptor;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.NoSuchElementException;
-import java.util.Optional;
+import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@Transactional
 public class ChessService {
 
     private static final int ROW_INDEX = 0;
     private static final int COLUMN_INDEX = 1;
 
-    private static final String NOT_HAVE_GAME = "해당하는 게임이 없습니다.";
-
     private final GameDao gameDao;
     private final PieceDao pieceDao;
-
-    private final Map<Long, ChessGame> chessGames;
 
     public ChessService(GameDao gameDao, PieceDao pieceDao) {
         this.gameDao = gameDao;
         this.pieceDao = pieceDao;
-        this.chessGames = new HashMap<>();
     }
 
-    public ChessGameResponse loadGame(long gameId) {
-        Optional<GameState> maybeGameState = gameDao.load(gameId);
-        GameState gameState = maybeGameState.orElseThrow(NoSuchElementException::new);
+    public Long createGame(CreateGameRequest createGameRequest) {
+        String salt = PasswordEncryptor.generateSalt();
+        GameEntity gameEntity = GameEntity.toSave(createGameRequest.getGameName(),
+                PasswordEncryptor.encrypt(createGameRequest.getPassword(), salt), salt,
+                GameState.READY);
+        ChessGame chessGame = new ChessGame(new Board(new CreateCompleteBoardStrategy()));
+        Long gameId = gameDao.save(gameEntity);
+        saveBoard(gameId, chessGame.getBoard());
+        return gameId;
+    }
+
+    public void saveBoard(Long gameId, Map<Position, Piece> board) {
+        board.forEach((position, piece) -> savePiece(gameId, position, piece));
+    }
+
+    private void savePiece(Long gameId, Position position, Piece piece) {
+        pieceDao.save(gameId, position, piece);
+    }
+
+    public ChessGamesResponse loadAllGames() {
+        return new ChessGamesResponse(gameDao.findAll()
+                .stream()
+                .map(entity -> new GameRoomResponse(entity.getId(), entity.getName()))
+                .collect(Collectors.toList()));
+    }
+
+    public ChessGameResponse loadGame(Long gameId) {
+        return new ChessGameResponse(createChessGameObject(gameId));
+    }
+
+    private ChessGame createChessGameObject(Long gameId) {
+        GameEntity gameEntity = gameDao.findById(gameId);
+        GameState gameState = gameEntity.getState();
         Board board = createBoard(gameId);
-        chessGames.put(gameId, new ChessGame(board, gameState));
-        return new ChessGameResponse(getChessGame(gameId));
+        return new ChessGame(board, gameState);
     }
 
     private Board createBoard(long gameId) {
@@ -58,68 +87,58 @@ public class ChessService {
         return new Board(() -> pieces);
     }
 
-    private ChessGame getChessGame(long gameId) {
-        if (!chessGames.containsKey(gameId)) {
-            throw new IllegalArgumentException(NOT_HAVE_GAME);
-        }
-        return chessGames.get(gameId);
+    public ChessGameResponse startGame(Long gameId) {
+        gameDao.updateState(gameId, GameState.WHITE_RUNNING);
+        return loadGame(gameId);
     }
 
-    public ChessGameResponse createGame(long gameId) {
-        ChessGame chessGame = new ChessGame(new Board(new CreateCompleteBoardStrategy()));
-        gameDao.save(gameId);
-        saveBoard(gameId, chessGame.getBoard());
-        chessGames.put(gameId, chessGame);
-        return new ChessGameResponse(chessGame);
+    public ChessGameResponse resetGame(Long gameId) {
+        pieceDao.deleteByGameId(gameId);
+        gameDao.updateState(gameId, GameState.READY);
+        Board board = new Board(new CreateCompleteBoardStrategy());
+        saveBoard(gameId, board.getBoard());
+
+        return new ChessGameResponse(new ChessGame(board));
     }
 
-    public void saveBoard(long gameId, Map<Position, Piece> board) {
-        board.forEach((position, piece) -> savePiece(gameId, position, piece));
-    }
-
-    public void savePiece(long gameId, Position position, Piece piece) {
-        if (pieceDao.find(gameId, position).isEmpty()) {
-            pieceDao.save(gameId, position, piece);
-        }
-    }
-
-    public ChessGameResponse startGame(long gameId) {
-        ChessGame chessGame = getChessGame(gameId);
-        chessGame.start();
-        gameDao.updateState(gameId, chessGame.getGameState());
-        return new ChessGameResponse(chessGame);
-    }
-
-    public ChessGameResponse resetGame(long gameId) {
-        gameDao.delete(gameId);
-        return createGame(gameId);
-    }
-
-    public ChessGameResponse move(long gameId, MoveRequest moveRequest) {
-        ChessGame chessGame = getChessGame(gameId);
+    public ChessGameResponse move(Long gameId, MoveRequest moveRequest) {
+        ChessGame chessGame = createChessGameObject(gameId);
         Position start = parseStringToPosition(moveRequest.getStart());
         Position target = parseStringToPosition(moveRequest.getTarget());
         chessGame.move(start, target);
         if (pieceDao.find(gameId, target).isPresent()) {
-            pieceDao.delete(gameId, target);
+            pieceDao.deleteByGameIdAndPosition(gameId, target);
         }
         pieceDao.updatePosition(gameId, start, target);
         gameDao.updateState(gameId, chessGame.getGameState());
         return new ChessGameResponse(chessGame);
     }
 
-    public StatusResponse status(long gameId) {
-        ChessGame chessGame = getChessGame(gameId);
+    public StatusResponse status(Long gameId) {
+        ChessGame chessGame = createChessGameObject(gameId);
         return new StatusResponse(chessGame.createStatus());
     }
 
-    public EndResponse endGame(long gameId) {
+    public void deleteGame(Long gameId, String password) {
+        GameEntity gameEntity = gameDao.findById(gameId);
+        authenticate(password, gameEntity);
+        validateDeletable(gameEntity);
         gameDao.delete(gameId);
-        return new EndResponse("게임이 종료되었습니다.");
     }
 
-    public void deleteGame(long gameId) {
-        gameDao.delete(gameId);
+    private void authenticate(String password, GameEntity gameEntity) {
+        String actualPassword = gameEntity.getPassword();
+        String salt = gameEntity.getSalt();
+        String sentPassword = PasswordEncryptor.encrypt(password, salt);
+        if (!sentPassword.equals(actualPassword)) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+    }
+
+    private void validateDeletable(GameEntity gameEntity) {
+        if (!gameEntity.getState().isFinished()) {
+            throw new IllegalArgumentException("게임이 종료되기 전에는 삭제할 수 없습니다.");
+        }
     }
 
     private Position parseStringToPosition(final String rawPosition) {
